API Reference

PartitionedArrays.PSparseMatrix — Type

PSparseMatrix(init, I, J, V, rows::PRange, cols::PRange, args...; ids, kwargs...)

Create a new PSparseMatrix from the COO-vectors I, J, and V using init as the initialization method (e.g. sparse for constructing SparseMatrixCSCs on every process, or sparsecsr for constructing SparseMatrixCSR on every process). I, J, and V, should all be AbstractPData-wrapped arrays containing the row-IDs, column-IDs, and values, respectively. rows and cols should be PRanges describing the process ownership of the rows and columns. If the IDs (in I, and J) are given in global-enumeration this should be specified by passing ids=:global as a keyword argument (the constructor will then internally renumber to local IDs), and if the IDs are given in process-local enumeration this should be specified by passing ids=:local.

PSparseMatrix(init, I, J, V, rows::Int, cols::Int, args...; kwargs...)

Same as above, but with rows and cols given as integers. This method requires IDs in I, and J to be globally enumerated (and thus passing ids=:global).

PSparseMatrix(I, J, V, args...; kwargs...)

Same as the methods above, except providing sparse as the default initialization method.

PartitionedArrays.async_assemble! — Function

async_assemble!(I, J, V, rows::PRange) -> Task

Create a Task for syncronizing the COO-vectors I, J, and V, i.e. sending the triplet (i, j, v) to the process that owns row i, The row ownership is specified by rows.

The Task that is returned is not scheduled. To execute the assembly use schedule and wait.

Example

assembly_task = async_assemble!(I, J, V, rows)
map_parts(schedule, assembly_task)
# Could do other work here while the task is finishing
map_parts(wait, assembly_task)

PartitionedArrays.assemble! — Function

assemble!(args...; kwargs...)

Blocking version of async_assemble!.

PartitionedArrays.add_gids! — Function

add_gids!(a::PRange, gids::AbstractPData, neighbors_snd=nothing, neighbors_rcv=nothing; kwargs...)

For each part, associate the global IDs in gids to the PRange a. Ghost IDs are added for the global IDs that are not owned by the corresponding part.

See also add_gids.

PartitionedArrays.map_parts — Function

map_parts(f, xs::AbstractPData...) -> AbstractPData

Transform the partitioned data in xs by applying f to the underlying local parts, resulting in another partitioned data. If f returns an iterable object, the resulting partitioned data will contain an iterable object at each part. This object can be unpacked into several partitioned data of a single local element.

Example

using PartitionedArrays
backend = SequentialBackend()
p = get_part_ids(backend,3)
x,y = map_parts(p) do p
    (sin(p),cos(p))
end
map_parts(x,p) do x,p
    @assert x == sin(p)
end
z = map_parts(+,x,y)
map_parts(p,z) do p,z
    @assert z == sin(p)+cos(p)
end