Benchmarking

PTimer

PartitionedArrays.PTimer — Type

struct PTimer{...}

Type used to benchmark distributed applications based on PartitionedArrays.

Properties

Properties and type parameters are private

Sub-type hierarchy

PTimer{...} <: Any

PartitionedArrays.PTimer — Method

PTimer(ranks;verbose::Bool=false)

Construct an instance of PTimer by using the same data distribution as for ranks. If verbose==true, then a message will be printed each time a new section is added in the timer when calling toc!.

Measuring elapsed times

PartitionedArrays.tic! — Function

tic!(t::PTimer;barrier=false)

Reset the timer t to start measuring the time in a section. If barrier==true, all process will be synchronized before resetting the timer if using a distributed back-end. For MPI, this will result in a call to MPI.Barrier.

PartitionedArrays.toc! — Function

toc!(t::PTimer,name::String)

Finish measuring a code section with name name in timer t.

Post-process

PartitionedArrays.statistics — Function

statistics(t::PTimer)

Return a dictionary with statistics of the compute time for the sections currently stored in the timer t.