PSparseMatrix

Type signature

PartitionedArrays.PSparseMatrix — Type

struct PSparseMatrix{V,A,B,C,...}

PSparseMatrix (partitioned sparse matrix) is a type representing a matrix whose rows are distributed (a.k.a. partitioned) over different parts for distributed-memory parallel computations. Each part stores a subset of the rows of the matrix and their corresponding non zero columns.

This type overloads numerous array-like operations with corresponding parallel implementations.

Properties

matrix_partition::A

row_partition::B

col_partition::C

matrix_partition[i] contains a (sparse) matrix with the local rows and the corresponding nonzero columns (the local columns) in the part number i. eltype(matrix_partition) == V. row_partition[i] and col_partition[i] contain information about the local, own, and ghost rows and columns respectively in part number i. The types eltype(row_partition) and eltype(col_partition) implement the AbstractLocalIndices interface.

The rest of fields of this struct and type parameters are private.

Supertype hierarchy

PSparseMatrix{V,A,B,C,...} <: AbstractMatrix{T}

with T=eltype(V).

Accessors

PartitionedArrays.local_values — Method

local_values(a::PSparseMatrix)

Get a vector of matrices containing the local rows and columns in each part of a.

PartitionedArrays.own_values — Method

own_values(a::PSparseMatrix)

Get a vector of matrices containing the own rows and columns in each part of a.

PartitionedArrays.ghost_values — Method

ghost_values(a::PSparseMatrix)

Get a vector of matrices containing the ghost rows and columns in each part of a.

PartitionedArrays.own_ghost_values — Method

own_ghost_values(a::PSparseMatrix)

Get a vector of matrices containing the own rows and ghost columns in each part of a.

PartitionedArrays.ghost_own_values — Method

ghost_own_values(a::PSparseMatrix)

Get a vector of matrices containing the ghost rows and own columns in each part of a.

Constructors

PartitionedArrays.PSparseMatrix — Method

PSparseMatrix(matrix_partition,row_partition,col_partition)

Build an instance for PSparseMatrix from the underlying fields matrix_partition, row_partition, and col_partition.

PartitionedArrays.psparse — Function

psparse(f,row_partition,col_partition)

Build an instance of PSparseMatrix from the initialization function f and the partition for rows and columns row_partition and col_partition.

Equivalent to

matrix_partition = map(f,row_partition,col_partition)
PSparseMatrix(matrix_partition,row_partition,col_partition)

PartitionedArrays.psparse! — Function

psparse!([f,]I,J,V,row_partition,col_partition;discover_rows=true) -> Task

Crate an instance of PSparseMatrix by setting arbitrary entries from each of the underlying parts. It returns a task that produces the instance of PSparseMatrix allowing latency hiding while performing the communications needed in its setup.

Assembly

PartitionedArrays.assemble! — Method

assemble!([op,] a::PSparseMatrix) -> Task

Transfer the ghost rows to their owner part and insert them according with the insertion operation op (+ by default). It returns a task that produces a with updated values. After the transfer, the source ghost rows are set to zero.